@waylan 2016-02-26
It would be interesting to see some benchmarks before and after this change. Mostly, I'm just curious how much of a difference it makes.

However, before it will be accepted, all the tests need to pass. Currently, [3 tests are failing][1] related to syntax highlighting with the CodeHilite extension. That extensions actually [stores its output in the HtmlStash][2] so I'm assuming this change is causing the highlighted code blocks to not be swapped in for their placeholder's properly. Unfortunately, due to inconsistencies in output from different versions of Pygments, our tests are a little strange (we use `startswith` to verify a Pygments HTML block was returned without actually checking the contents of the block) and the tests' error messages are less than helpful in identifying the problem.

However, it occurs to me that if your patch is relying on the placeholders being in a specific order (both in the HTMLStash and in the document), with the CodeHilite extension that order is not necessarily preserved between document and stash. Any such patch needs to support that possibility.

[1]: https://travis-ci.org/waylan/Python-Markdown/jobs/112102071#L250
[2]: https://github.com/waylan/Python-Markdown/blob/master/markdown/extensions/codehilite.py#L218