@mitya57 2016-02-26
Is there some error in the `HtmlOutput` plugin? The traceback from Python 3.x log looks like an internal testsuite error in that plugin.