@Griffon26 2016-02-26
I added an assertStartsWith function to the tests, so it now shows exactly why the assertions fail (at least for 2.7 and pypy):

```
AssertionError: u'<div class="codehilite"><pre><span class="hll"' not found at the start of u'<div class="codehilite"><pre><span></span><span cla...'
```

There's an additional `<span></span>` fragment in the the output. Any idea how that could happen at TravisCI but not on my or Dmitry's machines?