@waylan 2016-02-28
Groan. Yep, that would be it. Thanks for finding it @mitya57.