@Griffon26 2016-02-28
I've checked why the failure caused by pygments is not reported properly for python3 and I think it's [this issue](https://github.com/nose-devs/nose/issues/725). Unfortunately that got closed with "upgrade to nose2".

I do have a work-around (similar to what they are doing [here](https://github.com/ionelmc/nose-htmloutput/issues/1)):

    diff --git a/tests/plugins.py b/tests/plugins.py
    index 90c5c0d..4e7af97 100644
    --- a/tests/plugins.py
    +++ b/tests/plugins.py
    @@ -97,6 +97,8 @@ class HtmlOutput(Plugin):
 
         def formatErr(self, err):
             exctype, value, tb = err
    +        if not isinstance(value, exctype):
    +            value = exctype(value)
             return ''.join(traceback.format_exception(exctype, value, tb))
 
         def startContext(self, ctx):

If you'd like me to commit this, I'll add it to this pull request.

Is there anything else I can do?