@mitya57 2016-03-03
@waylan: ping? I can take care of this pull request myself if you prefer (i.e. make sure the tests pass and merge it).