@mitya57 2016-03-03
Merged. The tests pass everywhere according to Travis. Thanks @Griffon26 :)